docs: simplify foundation stories
Description bloat (verbose prose, bilingual sections, redundant tables)
across foundation pages. Trim to under 15 lines per meta.

- accent.stories removed (accent docs absorbed into colors)
- typography-extras removed (consolidated into typography)
- Other foundation pages: shorter meta description, leaner examples
